Less confusion regarding CL intro text (#7041)
# About the pull request

In a recent round there was confusion about when the CL can be given
legal orders and when the orders would not be legal. It was pointed out
that the CL intro text is a bit confusing on this.
https://cm-ss13.com/wiki/Rank is a page editable by wiki admins only,
like the Standard Operating Procedure or Marine Law pages. This
paraphrases the info from there on when civilians can be given legal
orders, and when they cannot.

# Explain why it's good for the game

Less confusion for CL if CL gets arrested for not following reasonable
orders of CO for example.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: CL intro text is more clear regarding legal orders
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>