Mapping: LV522 Changes to overall map, including the size, poplimit, chokepoints and other such things (#6977)
# About the pull request

This PR touches a lot of LV522 but the broad strokes of it are making
the map more fit for a smaller population by cutting out the "fat"
aswell as attempting to fix chokepoint-related issues in the reactor

I started this project about 2 months ago but burnout hit me like a
solid brick I've updated the map but there might be some minor bugs
carried over when I've copypasted everything over. I fixed as many as I
saw but you know how it is.

Aswell as all that I've added shortfog to the LZ tunnel so FORECON cant
get inside. This hold has been a deep-rooted meta for the map and we'd
like to see something different as it's a bit too easy

# Explain why it's good for the game

I think LV522 really struggles with its size a lot of space that doesn't
really needs to be there but was mapped in anyway for a shitty "Well
it's a highpop map" mentality I hope to trim away the fat get the map
more acceptable for sub 130 players and hopefully make the reactor more
fun to be inside of as a marine


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: SpartanBobby
maptweak: LV522 changes relating to the overall size of the map,
redefining chokepoints in the reactor, changes to engineering and the
west side of the map and removing the poplock
maptweak: LV522, LZ1 Tunnel (the place FORECON holds) has been given fog
for the start of the round meaning FORECON will have to adapt to
survive.
maptweak: buffed crashed ship lootspawns on LV522
/:cl: