[Solaris Ridge] Aesthetic Update V2 (#6602)
# About the pull request

Continuation of this PR: https://github.com/cmss13-devs/cmss13/pull/6152

--

Effects a significant aesthetic makeover over parts of Solaris Ridge.

This PR focuses more on the Northern half of the colony. I intend to
followup in the future with doing more work on the southern half and the
caves.

--

# Explain why it's good for the game

I feel a long term goal of mapping should be to enhance our older maps
to make them look far more "alive", lived in and aesthetically pleasing.
But keeping the spirit of the similar layout of the original map.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: The northern half of the main colony of Solaris Ridge has
received an aesthetic upgrade.

/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>