Adds a staff button to check for whitelisted players. (#7062)
# About the pull request

It's not uncommon for events to be run with whitelisted roles, and this
will make it much easier to check if someone with the right whitelist is
online.

# Changelog
:cl:
add: Added a staff event verb to check for a particular whitelist and
report online players that have it. Code for the proc from HarryOB
/:cl:

---------

Co-authored-by: kiVts <48099872+kiVts@users.noreply.github.com>