Updates close backpack icon (#6820)
# About the pull request

Replaces red cross made in 2007 with an actual icon.

Before:


![dreamseeker_ibtZ94zQ86](https://github.com/user-attachments/assets/e9c98d63-d163-45e8-9bf2-56962230f756)

After:


![dreamseeker_fB1ekGWiK7](https://github.com/user-attachments/assets/0cdc4df4-7ae6-437e-979a-e50f5abe291d)


# Explain why it's good for the game

Placeholder bad, but not only that, the cross itself is a hitbox for
click detection, and cross isn't really ideal shape for such icon since
most of the people would try to hit the centre of it (~81 pixels area),
and new icon has bigger centre area overall (~144 pixels) and is easier
to hit because it's a lot harder to miss on a big square without blind
spots and which is actually connected to the main UI body.


Sprite itself from
[TGMC](https://github.com/tgstation/TerraGov-Marine-Corps).



# Changelog
:cl:
imageadd: replaces ancient close backpack red cross with a modern icon
/:cl: