Adds a webhook for sending the top 10 overtiming procs to discord (#7139)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->
As the title says. Will list out the 10 most overtiming procs. Happens
at the end of a round.

Some important documentation here: Overtime is the time that a proc runs
over tick for. Any amount of overtime is typically bad and, if high
enough, needs to have a `CHECK_TICK` call within itself or it needs to
run on a subsystem.

High amounts of overtime directly links to any of the large lagspikes
you might see during gameplay. If you notice that everyone freezes for a
few seconds during gameplay, that's because a proc suddenly spiked in
overtime.

Incremental amounts of overtime can lead to gameplay stutters as well.
It's impossible to completely eliminate overtime but it provides a good
performance target for which procs have the most dangerous serverside
cost, depending on how much average overtime they typically have.

# Explain why it's good for the game
Might get more visibility on procs that actually cause performance
issues. Since profiler information is probably not easily available to
players and contributors, this might provide an optimization goal for
them, and a clear idea of what might be causing serverside lagspikes
during gameplay.

# Testing Photographs and Procedure

![image](https://github.com/user-attachments/assets/6ded10e6-19c2-4bb9-9770-dfc0b28edcdb)

# Changelog
:cl:
add: Added a webhook to send the top 10 most overtiming procs to a
discord webhook on server shutdown.
/:cl:

---------

Co-authored-by: Watermelon914 <3052169-Watermelon914@users.noreply.gitlab.com>