SG falloff rebalance and far sight removal (#7065)
# About the pull request
This PR:
1. Removes far-sight
2. Increases falloff to 4 (from 1)
3. Increases effective range to 7 (from 0) - This means falloff starts
counting at the 7th tile (edge of the screen)
4. Needs testing
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Getting hit by bullets offscreen as well as dealing no damage on your
own screen is rather annoying.
On a more serious note SGs are currently encouraged to go as close as
possible in order to actually deal damage properly which is in contrast
to the purpose of the role of providing suppressing fire from the back
(AP bullets for example losing around 50% of their damage at the edge of
the screen)
As for far-sight that was a rubber-band to try to fix the above (SGs
going melee range) but just ended up being annoying to play against.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Removed far-sight from SG sights
balance: Increased smartgun bullet falloff to 4 from 1
balance: Increased smartgun bullet effective range to 7 from 0
/:cl: