Bumps MW Fireman in line with rest of MPs (#7144)
# About the pull request
Changes Fireman from Trained to Skilled like the rest of the MPs and
some Command Staff.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Presumably a previous oversight.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: 
balance: Buffs MW Fireman skill in line with the rest of MPs
/:cl: