Adds flare packs to FTL vendors  (#7118)
# About the pull request

Adds signal flare packs to the FTL vendors for 5 points a pack. 

# Explain why it's good for the game

FTL is a role that's based around leadership and calling fire support,
in practice. It seems silly they don't have the option to buy more
signal flares to do this with.

# Testing Photographs and Procedure
Booted the game, signal flare packs in vendor, each costs 5 points. 
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
add: FTL vendors get signal flare packs for 5 points.
/:cl: