Injector and Coagulator plates now function (#7173)
# About the pull request
made the plates actually work.
Fixes  #7142

All code was made by kiVts, I just took the code and made a PR for them
now because #7023 is on draft until it's done.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

The things that research spends biomass points on should work???
# Testing Photographs and Procedure
tested on a private server, stabbed myself a few times with a knife, no
bloodloss, injector plate works (with all OD modes functional).


# Changelog
:cl: Private-Tristan, KiVts
fix: Injector and Coagulator plates are now functional.
/:cl: