bracer buff for predators (#7196)
# About the pull request

Increases the recharge rate for the Yautja bracer

# Explain why it's good for the game

It allows Yautjas to actually play the game more, bracers are very
rarely used for combat if ever in a round. I don't expect my
whitelistees to AFK for 10 minutes because they can't use the NVGs to
play the game anymore.

# Testing Photographs and Procedure

numberchange


# Changelog

:cl: stalkerino
balance:  increased recharge for pred bracer 
/:cl: