MPs now latejoin in the Brig (#7162)
# About the pull request

MPs now latejoin in the Brig as opposed to cryo.

# Explain why it's good for the game

Alot of other shipside roles have dedicated latejoin spots, (and most
others like Marines having ones close to prep) and having one as MP in
cryo, more than anything is a hassle as you have to run across the
*entire* ship just to gear up, esp on Red Alert.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: MPs now latejoin in the Brig
/:cl: