Adds a preference to show your tag as Queen (#7190)
# About the pull request
Adds a preference to show your tag as queen
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Initially this was added to protect queen players however some people
don't mind it and actually want to show off their name and they should
be able to do that.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: Added a preference to show your tag as Queen, turned off by
default.
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>