Fix runtime with preview dummy related to signals (#7219)
- Preview dummy will throw a stack_trace if a trait applies any signals
since it applies all traits on update without clearing them
# About the pull request

Fixed a potential runtime for any character traits that want to register
signals.

Also some minor formatting changes.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Fix runtime
# Testing Photographs and Procedure
Tested with a signal-adding trait I was experimenting with

# Changelog
This is just a fix for a potential issue so not playerfacing