Adds a Bomb Suit/EOD Locker to the MP Armory (#7176)
# About the pull request

Replaces one of the Riot Lockers in the Armory with one EOD Locker.
(Same one found in OT Room but the Sec variant)

# Explain why it's good for the game

As it stands, if MPs, or even the ship receives a bomb threat, the only
relatively safe way of dealing with it is breaking into OT for their
Bomb Suit Locker. MPs should have their own way of dealing with bomb
threats without taking things from unrelated crew.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added an EOD Locker in the MP Armory.
/:cl: