Allows using empty verb on storage box items (#7265)
# About the pull request
This makes it so that the empty verb can be used on storage box items to
empty them as a quality of life thing. This impacts all types of objects
that use STORAGE_FLAGS_BOX, specifically:

- cardboard boxes including pre-filled ones
- pill bottles
- medical kits
- pill packets
- engineering kits

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Being able to use the empty verb rather than taking items out one at a
time is just a QOL thing and shouldn't impact balance in any way. If I
time myself taking items out of the box 1 by 1 and dropping it on the
floor it takes nearly the same amount of time as the delay that is
imposed by the empty verb. It just makes it so you don't have to click
rapidly to accomplish it.

The fact that XOR was being used on this flag is also a bit strange to
me. Both operands (default storage flags) and allow storage drawing were
1s which resulted in a 0 in that position. With the way this was
written, if the DEFAULT_STORAGE_FLAGS were ever to be set so that
emptying was disabled, this would then ENABLE it for boxes but disable
it for everything else which doesn't make a lot of sense.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: boxes can be emptied using empty verb
/:cl: