adds a containment cell to the Wey-Yu ERT station (#7257)
# About the pull request

It's always seemed idiosyncratic to me that the CL has a secured
containment cell for specimens, and one of the 'classic' forms of CL
shenanigans is securing differing alien samples, yet the station these
ERTs return to has no containment cell for them.

it seemed like a pretty easy change to dip my toes into CM coding, so I
figured I'd go for it.

also added medlinks to the vendors on the station, while I'm updating
it, since that seems like just an oversight with the medlink PR(or an
unwillingness to touch like 40 map files at once to add them to all the
random maps.)

# Explain why it's good for the game

Gives Wey-Yu ERTs and event characters more space to roleplay with the
specimens they are often attempting to capture.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Map editor preview

![image](https://github.com/user-attachments/assets/93e56c81-609a-4d9f-9614-63685d1d9120)

In-game preview

![image](https://github.com/user-attachments/assets/6656f518-8f6f-4e27-b21a-c8e36169aa5b)

In-character preview

![image](https://github.com/user-attachments/assets/51af8689-6ea7-4ec6-ac5c-1539a093218f)

</details>


# Changelog

:cl: detectivegoogle
maptweak: added a containment cell to the Wey-Yu ERT station
/:cl: