Ventcrawl notification fix (#7258)
# About the pull request
Fixes exiting through a pipe segment that's hidden underneath a floor
tile not sending a warning message to nearby mobs.
This allowed for practically unreactable surprise attacks, namely from
huggers.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Bugs are bad for the game, especially if they are conveniently (and
unknowingly) exploitable
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

https://streamable.com/qk3b3s

</details>


# Changelog
:cl:
fix: Fixed exiting ventcrawling through a covered pipe segment not
triggering a warning notification for nearby mobs
/:cl: