Adds Scandinavian as a language option in the traits menu (#7247)
# About the pull request

I have added Scandinavian as a language in the traits option. The
syllables used are mostly the same as the German ones after being
translated into Norwegian, I chose Norwegian since that's the one I'm
most familiar with. Scandinavian languages are lumped together because:
A: They are mutually intelligible.
B: They are hard to distinguish if you don't speak it.
C: I assume with time the languages have only gotten more similar.
D: It gives an option for characters from multiple countries to have a
language.

NOTE: The key for the language is currently set at 0, as all other
numbers are taken. I do not like this, I would rather it be set to
something else, however I do not know all the keys, if there is a better
suggestion feel free to change it.

# Explain why it's good for the game

Allows characters more roleplay options, while German exists and is
similar to Scandinavian languages, they are distinct and it does
restrict roleplay options. I generally believe that having more options
to flesh out a character is good, especially when it doesn't impact the
game itself majorly.


# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>


https://github.com/user-attachments/assets/6c771fd0-6fcf-40aa-85a9-f597b9f2abdb

</details>


# Changelog

:cl: lokideslimee
add: Added Scandinavian as a language
code: Coded in the Scandinavian
code: Added Scandinavian as a language for mobs that should know the
language
/:cl:

---------

Co-authored-by: LachlanSyn <a1885472@adelaide.edu.au>