Fix's the Inconsistencies with the joining a role message and the Chain of Command (#7218)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request
With the new change to marine law Doctors, Surgeons, and Researchers are
now part of the chain of command. This PR changes the intro message so
that it does not say they are not part of the chain of command.


# Explain why it's good for the game
Cleans up inconsistencies in the code from what the Chain of Command
says.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
label your changes in the changelog. Please note that maintainers freely
reserve the right to remove and add tags should they deem it
appropriate. You can attempt to finagle the system all you want, but
it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Maintainers freely reserve the right to remove and add
tags should they deem it appropriate. -->

:cl: Lukas292001
spellcheck: Changes doctor and researcher join description, from saying
they are not in the chain of command to saying they are.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->