Changes /area/yautja checks to look for a flag (#7255)
# About the pull request

Stops the few places that typecheck for /area/yautja to look for a
flags_area flag instead, this will allow more flexibility in the future,
and gets rid of some loathesome typechecking.

# Explain why it's good for the game

Future flexibility. Less typechecking.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
code: Yautja gear claim and tracking now checks for an area flag instead
of area typechecking.
/:cl: