Abominations (Predaliens) can no longer be gibbed (#7253)
# About the pull request

See title. the code is the same that prevents queen from gibbing.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Killing a queen is a p. big feat for marines, and marines get to parade
the body around or send it to research for points, but abominations
don't have the same gib protection despite being super rare to see in a
round, and killing it just for it to gib instead of leaving a corpse
kinda sucks.
# Testing Photographs and Procedure

Tested and works 


# Changelog

:cl:
qol: Abominations (Predaliens) are now gib immune
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>