Fixes Queen being unable to stomp xenos (#7275)
# About the pull request

Queen was unable to stomp *any* other xeno after
https://github.com/cmss13-devs/cmss13/pull/7228

We want to OR the two conditions first, and then AND them. Currently we
were AND'ing the first 2 parameters (giving 0 in most cases), and then
OR'ing them. (giving non-zero)

Adds a continue instead of a return in the case that we have multiple
xenos on the tile, some of which we want to stomp

Fixed now.

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Fixes being unable to stomp xenos as Queen
/:cl: