Fusion reactor oversight fix (#7157)
# About the pull request
fixes #7127


This if/then/otherwise statement was activating the wrong message when
it was overloaded meaning that when you were restoring safety’s then
chat would say that you weren’t. switched the then statements around and
now it’s working again
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: fusion reactor overload messages now show correctly
/:cl:

---------

Co-authored-by: BeagleGaming1 <56142455+BeagleGaming1@users.noreply.github.com>