[Pred] Adds a no block alternative to combistick (#7268)
# About the pull request

This pr adds in the War axe, as an alternative to the combistick. This
weapon has no block chance (as seen below), and is intended to be a
sidegrade from the combi. In return, it is a one handed weapon. Thus,
the combi is left as is, and will continue to be an option for future
hunts.

# Explain why it's good for the game

People really dislike the 30% block chance that comes associated with
the combi. Landing shots or slashes against an already fast, tanky
antagonist but having the game tell you "no" is not fun or intuitive.
This PR is intended for the people who enjoy the throwing and tethered
aspect of the combi, but don't want to guarantee their victory against
anything short of a marine mob or a hive offensive.

# Testing Photographs and Procedure
Code was tested and ensured that it worked. Combi functions as normal,
and war axe functions as a correct copy of it (you are not able to "fold
down" the war axe, like you are with the combi).
Sprites are courtesy of Thwomper; huge shoutout to you for helping me
through my first project.


![image](https://github.com/user-attachments/assets/0d09cb85-5787-40a5-983b-8322dc4dbcd4)

<details>
<summary>Screenshots & Videos</summary>

How to choose it from the claim gear verb


https://github.com/user-attachments/assets/2fca01ee-b88b-4756-9bb4-6d8800d021c7

Functions like a combi. Yank Combi-stick is now called Yank Weapon, but
still uses the internal guts of combi keybinding (this will not break
your keybinds)


https://github.com/user-attachments/assets/e2d8f307-e383-4405-aeb3-bc077feb3f1b

Reference of Combistick blocking


https://github.com/user-attachments/assets/1a666d8f-e943-4153-8f33-9053bd033bfc

Showcase of War axe having 0 block chance


https://github.com/user-attachments/assets/8f9612b4-3925-4c68-840e-dca8467c85bc

Thralls are still able to pick the combi. They do not have an option for
the alt version (similar to the other alt weapons)


https://github.com/user-attachments/assets/f840f866-fd85-4e1c-954f-cd404c73ec06


</details>


# Changelog

:cl:
add: Added a varient of combistick that has no block chance.
/:cl:

---------

Co-authored-by: Blundir <100090741+Blundir@users.noreply.github.com>