Hostile ERT clarity messages (#7302)
# About the pull request

Makes it VERY clear to all ERT members if they're hostile or not.
Also reminds staff in the finalisation log if hostile or not.
Some ERTs (like MARSOC and special event ERTs) have not had messages
added.

# Explain why it's good for the game

Peace of mind and no excuses.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added a giant message to end of ERT spawns so they know if they're
hostile or not.
add: CLF armor now protects from cold weather so hostile survivors are
not penalized.
/:cl: