Increases the amount of mk2/m39 mags in preps (#7303)
# About the pull request
When I made #7153 I decreased the starting mk2 and m39 mags too much, so
i'm increasing them again
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
I want more mk2 mags in prep vendor so that I don't need to loot
multiple vendors for ammo.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Increased the starting amount of mk2 and m39 mags in prep
vendors
/:cl: